+++ ClientApp/src/App.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { Route } from 'react-router';
 import { Layout } from './components/Layout';
-import { Home } from './components/Home';
-import { Leagues } from './components/Leagues';
+import { Home } from './components/home-page/Home';
+import { Leagues } from './components/leagues-page/Leagues';
 
 export default class App extends Component {
   displayName = App.name
