+++ ClientApp/src/components/NavMenu.js
@@ -12,17 +12,12 @@ export class NavMenu extends Component {
       <Navbar inverse fixedTop fluid collapseOnSelect>
         <Navbar.Header>
           <Navbar.Brand>
-            <Link to={'/'}>football_history</Link>
+            <Link to={'/'}>Football History</Link>
           </Navbar.Brand>
           <Navbar.Toggle />
         </Navbar.Header>
         <Navbar.Collapse>
           <Nav>
-            <LinkContainer to={'/'} exact>
-              <NavItem>
-                <Glyphicon glyph='home' /> Home
-              </NavItem>
-            </LinkContainer>
             <LinkContainer to={'/leagues'}>
               <NavItem>
                 <Glyphicon glyph='th-list' /> Leagues
